fix(schematic): stop silently discarding coordinates and CDF meaning
Two defects from the 2026-09-10 tooling audit, both of the same shape: the
tool had the information and threw it away without saying so.

Defect B — coordinates were truncated to 0 and the call answered ok.
`json_i64_or(params.get("x"), 0)` gave `as_i64().unwrap_or(0)`, so both
`{"x": -1.5}` and `{"x": "7"}` placed the instance at the origin with
`status: ok`. The schematic grid is 0.0625; most real placements are not
integers. Coordinates are now f64 through dispatcher, ops, commands, editor,
CLI and batch spec — `dbCreateInst`'s `l_point` is in user units (IC23.1
skdfref.fnd) — and a present-but-non-numeric value is an error.

`move_instance` was already f64, so the two halves of "place it, then tidy
it" disagreed about what a coordinate is.

Same family, found while fixing it: `schematic.wire` parsed its points with
two `filter_map`s, so one mistyped coordinate silently shortened the wire
and still reported ok. Every point must now parse.

Live on IC23.1: place x=-1.5 y=4.0625 reads back as -1.5 / 4.0625;
x="7" errors instead of placing at the origin.

list_cdf_params — names without meaning.
It returned {name, type, value}, so a caller facing vsin's 135 names could
not tell `va` from `vaDBm` from `acm`. That is the gap that produced
`ampl=5m` on a vsin and a day chasing a flat transient. The CDF already
carries the answer, measured rather than assumed:

  prompt      135/135 on vsin, and on the PDK (`w` -> "Total Width")
  units       voltage / frequency / lengthMetric
  defValue    present; `w` defaults 300n while the instance holds 20u
  choices     1/135 (filenums), emitted as a JSON array when present
  description 0/135 — so prose still comes from analoglib.info

This is the only route that covers PDK devices, which no Cadence manual
documents. `simM = iPar("m")` survives the JSON escaping intact.

cargo test --release: 3203 passed / 0 failed / 7 ignored (was 3188/0/7).

(cherry picked from commit bc520d3e53460ee6c7452f1aa3bbb4a8d136a615)

+++ src/commands/schematic.rs
@@ -53,8 +53,8 @@ pub fn open(lib: &str, cell: &str, view: &str) -> Result<Value> {
 pub fn place(
     master: &str,
     name: &str,
-    x: i64,
-    y: i64,
+    x: f64,
+    y: f64,
     orient: Orient,
     params: &[(String, String)],
 ) -> Result<Value> {
@@ -76,7 +76,7 @@ pub fn place(
 }
 
 pub fn wire_from_strings(net: &str, points: &[String]) -> Result<Value> {
-    let pts: Vec<(i64, i64)> = points
+    let pts: Vec<(f64, f64)> = points
         .iter()
         .map(|s| {
             let (x, y) = s
@@ -93,7 +93,7 @@ pub fn wire_from_strings(net: &str, points: &[String]) -> Result<Value> {
     wire(net, &pts)
 }
 
-pub fn wire(net: &str, points: &[(i64, i64)]) -> Result<Value> {
+pub fn wire(net: &str, points: &[(f64, f64)]) -> Result<Value> {
     let client = VirtuosoClient::from_env()?;
     let skill = client.schematic.create_wire(points, "wire", net);
     let r = client.execute_skill(&skill, None)?;
@@ -122,7 +122,7 @@ pub fn conn(net: &str, from: &str, to: &str) -> Result<Value> {
     }))
 }
 
-pub fn label(net: &str, x: i64, y: i64) -> Result<Value> {
+pub fn label(net: &str, x: f64, y: f64) -> Result<Value> {
     let client = VirtuosoClient::from_env()?;
     let skill = client.schematic.create_wire_label(net, (x, y));
     let r = client.execute_skill(&skill, None)?;
@@ -132,7 +132,7 @@ pub fn label(net: &str, x: i64, y: i64) -> Result<Value> {
     }))
 }
 
-pub fn pin(net: &str, pin_type: &str, x: i64, y: i64) -> Result<Value> {
+pub fn pin(net: &str, pin_type: &str, x: f64, y: f64) -> Result<Value> {
     let client = VirtuosoClient::from_env()?;
     let skill = client.schematic.create_pin(net, pin_type, (x, y));
     let r = client.execute_skill(&skill, None)?;
@@ -169,8 +169,8 @@ pub fn save() -> Result<Value> {
 /// cosmetic: "default" (0.0625, centerCenter) or "clean" (0.125, lowerCenter)
 pub fn net_stub(
     net: &str,
-    x: i64,
-    y: i64,
+    x: f64,
+    y: f64,
     direction: &str,
     length: f64,
     cosmetic: &str,
@@ -255,9 +255,9 @@ pub struct SpecInstance {
     pub name: String,
     pub master: String, // "lib/cell"
     #[serde(default)]
-    pub x: i64,
+    pub x: f64,
     #[serde(default)]
-    pub y: i64,
+    pub y: f64,
     #[serde(default = "default_orient")]
     pub orient: Orient,
     #[serde(default)]
@@ -290,9 +290,9 @@ pub struct SpecPin {
     #[allow(dead_code)]
     pub connect: Option<String>, // "M2:G"
     #[serde(default)]
-    pub x: i64,
+    pub x: f64,
     #[serde(default)]
-    pub y: i64,
+    pub y: f64,
 }
 
 pub fn build(spec_path: &str) -> Result<Value> {
